% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{ch.zeroes.design1}
\alias{ch.zeroes.design1}
\title{Systematic checkerboard design.}
\usage{
ch.zeroes.design1(rates, reps, ydim = 100, width = 3)
}
\arguments{
\item{rates}{vector of rates.}

\item{reps}{number of repetitions.}

\item{ydim}{y-dimension of simulated trial design.}

\item{width}{width of strips.}

\item{length}{length of the checkeboard squares.}
}
\value{
Raster.
}
\description{
Systematic checkerboard design for a fertiliser trial, with multiple repetitions of Nil (rate[1])
and repetition of rates across strips (west to east).
}
\author{
Fiona Evans
}
