% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{incdec.design}
\alias{incdec.design}
\title{Systematic strip design.}
\usage{
incdec.design(rates, reps, length = 100, width = 3)
}
\arguments{
\item{rates}{vector of rates.}

\item{reps}{number of repetitions.}

\item{length}{y-dimension of simulated trial design.}

\item{width}{width of strips.}
}
\value{
Raster.
}
\description{
Systematic strip design for a fertiliser trial with increasing rates from west to east,
then decreasing.
}
\author{
Fiona Evans
}
