% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{incr.design}
\alias{incr.design}
\title{Systematic strip design.}
\usage{
incr.design(rates, reps, ydim = 100, width = 3)
}
\arguments{
\item{rates}{vector of rates.}

\item{reps}{number of repetitions.}

\item{ydim}{y-dimension of simulated trial design.}

\item{width}{width of strips.}
}
\value{
Raster.
}
\description{
Systematic strip design for a fertiliser trial with increasing rates from west to east.
}
\author{
Fiona Evans
}
