% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{spatial.sim}
\alias{spatial.sim}
\title{Simulate spatial data with spatial variability over a grid}
\usage{
spatial.sim(mean = 3, psill = 0.25, range = 30, nsim = 1,
  xdim = 100, ydim = 100, output.type = "RasterStack")
}
\arguments{
\item{mean}{Mean of the spatial data.}

\item{psill}{(partial) sill of the variogram model.}

\item{range}{range parameter of the variogram model.}

\item{nsim}{number of simulations.}

\item{xdim}{x-dimension of simulated yield (width).}

\item{ydim}{y-dimension of simulated yield (length).}

\item{output.type}{"RasterStack" / "SpatialPixelsDataFrame"}
}
\value{
RasterStack / SpatialPixelsDataFrame.
}
\description{
Simulate a spatial data with specified mean yield and spatial variability simulated  
as a spatially correlated random field predicted on a grid. 
The spatial correlation structure is defined by an exponential variogram model with
specified psill and range, using gstat function 'vgm' with model = "Exp". Higher values 
of psill will result in a wider range fo simulated yield values. Higher values of
range will result in larger autocorrrelated regions. Output can be specified as RasterStack or 
SpatialPixelsDataFrame.
}
\author{
Fiona Evans
}
