% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{sys.design}
\alias{sys.design}
\title{Systematic repeated design (lowest, up one, highest, down one) 
Systematic repeated strip design.}
\usage{
sys.design(rates, reps, length = 100, width = 3)
}
\arguments{
\item{rates}{vector of rates.}

\item{reps}{number of repetitions.}

\item{length}{y-dimension of simulated trial design.}

\item{width}{width of strips.}
}
\value{
Raster.
}
\description{
Systematic strip design for a fertiliser trial with the lowest rate in the west,
increasing to the next highest, then the highest rate, decreasing to second highest etc.,
repated exactly for each rep. Note the number of rates must be a a multiple of 4!!!
}
\author{
Fiona Evans
}
