% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{ch.rcbd.design}
\alias{ch.rcbd.design}
\title{Randomised complete block (RCBD) checkerboard design.}
\usage{
ch.rcbd.design(rates, reps, length, ydim = 100, width = 3)
}
\arguments{
\item{rates}{vector of rates.}

\item{reps}{number of repetitions.}

\item{length}{length of the checkeboard squares.}

\item{ydim}{y-dimension of simulated trial design.}

\item{width}{width of strips.}
}
\value{
Raster.
}
\description{
Randomised complete block (RCBD) checkerboard design for a fertiliser trial.
}
\author{
Fiona Evans
}
