% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{ch.zeroes.design2}
\alias{ch.zeroes.design2}
\title{Systematic checkerboard design.}
\usage{
ch.zeroes.design2(rates, reps, check.length, length = 100, width = 3)
}
\arguments{
\item{rates}{vector of rates.}

\item{reps}{number of repetitions.}

\item{check.length}{length of the checkerboard squares.}

\item{length}{y-dimension of simulated trial design}

\item{width}{width of strips}
}
\value{
Raster.
}
\description{
Systematic checkerboard design for a fertiliser trial, with multiple repetitions of Nil (rates[1])
and same rate repeated across strips (west to east).
}
\author{
Fiona Evans
}
