% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{rcbd.design}
\alias{rcbd.design}
\title{Randomised complete block (RCBD) strip design.}
\usage{
rcbd.design(rates, reps, ydim = 100, width = 3)
}
\arguments{
\item{rates}{vector of rates.}

\item{reps}{number of repetitions.}

\item{ydim}{y-dimension of simulated trial design.}

\item{width}{width of strips.}
}
\value{
Raster.
}
\description{
Randomised complete block (RCBD) strip design for a fertiliser trial.
}
\author{
Fiona Evans
}
